Columns cell Render

Columns cell Render

Usage

ag_col_render(
  x,
  columns = NULL,
  cellStyle = NULL,
  cellClass = NULL,
  cellClassRules = NULL,
  cellRenderer = NULL,
  cellRendererParams = NULL,
  cellRendererSelector = NULL,
  autoHeight = NULL,
  wrapText = NULL,
  enableCellChangeFlash = NULL,
  suppressCellFlash = NULL,
  ...
)

Arguments

cellStyle

An object of css values / or function returning an object of css values for a particular cell. More details

cellClass

Class to use for the cell. Can be string, array of strings, or function that returns a string or array of strings. More details

cellClassRules

Rules which can be applied to include certain CSS classes. More details

cellRenderer

Provide your own cell Renderer component for this column's cells.See: Cell Renderer

cellRendererParams

Params to be passed to the cellRenderer component.

cellRendererSelector

Callback to select which cell renderer to be used for a given row within the same column. More details

autoHeight

Set to true to have the grid calculate the height of a row based on contents of this column.

wrapText

Set to true to have the text wrap inside the cell - typically used with autoHeight.

enableCellChangeFlash

Set to true to flash a cell when it's refreshed.

suppressCellFlash

Set to true to prevent this column from flashing on changes. Only applicable if cell flashing is turned on for the grid.